+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2017-07-27
 - Get drives space infos
+- Get drive name from first line of configuration file
 
 ## 2017-07-26
 - Fixed bug on read only shares UnauthorizedAccessException trying to set file attributes and can't open a file